Refactored the whole OrdersWriter with an actual JSONObject package
This class creates a JSON object manually, which is not ideal. So we use the org.json library to create JSON objects, which simplifies the construction of the JSON string and makes the code cleaner and easier to read. We also have broken down the getContents method into smaller, more manageable methods (createOrderObject and createProductObject) that each handle a specific part of the JSON string creation. This makes the code easier to understand and maintain, addressing the "Long Method" code smell.