Put the toJSON methods in Order and Product classes
This follows the principle of object-oriented programming that states "objects should be responsible for their own behavior". This will help to keep our classes encapsulated and will make our code easier to maintain and understand.

The Order and Product classes are more flexible and reusable, because they can now provide a JSON representation of themselves independently of the OrdersWriter class.